+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Add "/magic clean" command to remove unowned lost wands.
  - Add owner to wand description, if no other description is present.
+ - Add owner filtering to "/magic list" command.
 
 ## 2.8.9